+++ lib/ui/events2/Event2DetailPanel.dart
@@ -391,22 +391,22 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
   ] : null;
 
   List<Widget>? get _priceDetailWidget{
-    bool isFree = _event?.free ?? false;
-    // "panel.event2.detail.general.free.title": "Free"
-    String priceText =isFree? "Free" : (_event?.cost ?? "Free");
-    String? additionalDescription = isFree? _event?.cost : null;
-    List<Widget>? details = priceText.isNotEmpty ? <Widget>[
-      _buildTextDetailWidget(priceText, 'cost'),
-    ] : null;
-    
-    if(details != null && StringUtils.isNotEmpty(additionalDescription)){
-      details.add(Container(padding: EdgeInsets.only(left: 28), child:
-        Row(children: [Expanded(child:
-            Text(additionalDescription!, style: Styles().textStyles?.getTextStyle("widget.item.regular"))),
-          ])));
+    List<Widget>? details = <Widget>[];
+    if (_event?.free != false) {
+      details.add(_buildTextDetailWidget(Localization().getStringEx('panel.event2.detail.general.free.title', 'Free'), 'cost'));
+      if (StringUtils.isNotEmpty(_event?.cost)) {
+        details.add(_buildTextDetailWidget(_event?.cost ?? '', 'cost',
+          textStyle: Styles().textStyles?.getTextStyle('widget.info.regular.thin'),
+          iconVisible: false,
+          maxLines: 2,
+          detailPadding: EdgeInsets.zero
+        ));
+      }
     }
-    details?.add( _detailSpacerWidget);
-
+    else if (StringUtils.isNotEmpty(_event?.cost)) {
+      details.add(_buildTextDetailWidget(_event?.cost ?? '', 'cost'));
+    }
+    details.add( _detailSpacerWidget);
     return details;
   }
 
@@ -727,15 +727,16 @@ class _Event2DetailPanelState extends State<Event2DetailPanel> implements Notifi
 
   Widget _buildTextDetailWidget(String text, String iconKey, {
     TextStyle? textStyle, // 'widget.info.medium' : 'widget.info.medium.underline'
+    int? maxLines = 1, TextOverflow? overflow = TextOverflow.ellipsis,
     EdgeInsetsGeometry detailPadding = const EdgeInsets.only(top: 4),
     EdgeInsetsGeometry iconPadding = const EdgeInsets.only(right: 6, top: 2, bottom: 2),
-    bool iconVisible = true, bool underlined = false, int maxLines = 1,
+    bool iconVisible = true, bool underlined = false,
   }) =>
     _buildDetailWidget(
       Text(text,
         style: textStyle ?? Styles().textStyles?.getTextStyle(underlined ? 'widget.info.medium.underline' : 'widget.info.medium'),
         maxLines: maxLines,
-        overflow: TextOverflow.ellipsis,
+        overflow: overflow,
       ),
       iconKey,
       detailPadding: detailPadding,