+++ lib/ui/home/HomeAthleticsGameDayWidget.dart
@@ -113,7 +113,7 @@ class _HomeAthleticsGameDayWidgetState extends State<HomeAthleticsGameDayWidget>
       return HomeSlantWidget(favoriteId: widget.favoriteId,
         title: Localization().getStringEx('widget.game_day.label.its_game_day', 'It\'s Game Day!'),
         titleIconKey: 'athletics',
-        child: HomeMessageCard(message: Localization().getStringEx('widget.game_day.label.no_games', 'No today games.')),
+        child: HomeMessageCard(message: Localization().getStringEx('widget.game_day.label.no_games', 'No Illini Big 10 events today.')),
       );
     }
   }