+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Fixed
 - Cannot update survey created in admin app [#3542](https://github.com/rokwire/illinois-app/issues/3542).
+- Displaying athletics events [#3544](https://github.com/rokwire/illinois-app/issues/3544).
 ### Changed
 - Enable cost description for free events [#3461](https://github.com/rokwire/illinois-app/issues/3461).